--- lib/ideromanizer.rb
@@ -1,5 +1,0 @@
-module IDeromanizer
-  def deromanize(roman)
-    raise NotImplementedError
-  end
-end
