--- lib/iromanizer.rb
@@ -1,5 +1,0 @@
-module IRomanizer
-  def romanize(arabian)
-    raise NotImplementedError
-  end
-end
