--- lib/roman_symbol.rb
@@ -1,7 +1,0 @@
-module RomanSymbol
-  ROMAN_SYMBOLS = ["I", "V", "X", "L", "C", "D", "M"]
-
-  def self.to_symbol(id)
-    return ROMAN_SYMBOLS[id]
-  end
-end
